+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+### Added
+
 - Added icons for dark and light themes for .bri files
 - Added Brioche LSP server status to the status bar
 - Added Brioche build command
@@ -9,6 +11,10 @@
 - Environment variables for LSP server via VSCode settings
 - Log level for LSP server via VSCode settings
 - Brioche binary path for LSP server via VSCode settings
+
+### Bug Fixes
+
+- Fixed Brioche LSP output channel from creating multiple instances
 
 ## [v0.2.0] - 2024-11-17